260421-restaurant-bot/pyproject.toml
[project]
name = "260421-restaurant-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "duckduckgo-search>=8.1.1",
    "graphviz>=0.21",
    "ollama>=0.6.1",
    "openai-agents[viz]>=0.14.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
